More SODA referee pushback
Some of this is pushback against this comment:

"Also, the authors state in the conclusions that B-trees (or a-b-trees)
are supposed to work particularly well under memory hierarchy
considerations. Yes, the explicitly cache-oblivious versions may be
difficult to implement, but certainly not a plain vanilla 2-4 tress. So
I find it strange that it is not included in the competition."